style: 스토리 파일 폴더 분리
