+++ src/stories/components/common/Button.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import Button from '../components/common/Button';
+import Button from '../../../components/common/Button';
 
 const meta: Meta<typeof Button> = {
   title: 'components/common/Button',
