+++ src/stories/components/common/Input.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import Input from '../components/common/Input';
+import Input from '../../../components/common/Input';
 
 const meta: Meta<typeof Input> = {
   title: 'components/common/Input',
@@ -51,6 +51,6 @@ export const UnderlinedAlignCenter: Story = {
   args: {
     ...DEFAULT_ARGS,
     underlined: true,
-    align: 'center'
+    align: 'center',
   },
 };
