+++ src/stories/components/common/Tooltip.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import Tooltip from '../components/common/Tooltip';
+import Tooltip from '../../../components/common/Tooltip';
 
 const meta: Meta<typeof Tooltip> = {
   title: 'components/common/Tooltip',
