+++ src/stories/pages/CardNameRegisterPage.stories.tsx
@@ -1,13 +1,15 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { BrowserRouter } from 'react-router-dom';
 
-import CardRegisterPage from '../pages/CardRegisterPage';
-import { CardFormValueContext } from '../contexts/CardFormContext';
-import { CardData, CardInfo } from '../types/card';
+import CardNameRegisterPage from '../../pages/CardNameRegisterPage';
+import CardFormProvider, {
+  CardFormValueContext,
+} from '../../contexts/CardFormContext';
+import { CardInfo } from '../../types/card';
 
-const meta: Meta<typeof CardRegisterPage> = {
-  title: 'pages/CardRegisterPage',
-  component: CardRegisterPage,
+const meta: Meta<typeof CardNameRegisterPage> = {
+  title: 'pages/CardNameRegisterPage',
+  component: CardNameRegisterPage,
   decorators: [
     (Story) => (
       <div
@@ -19,9 +21,11 @@ const meta: Meta<typeof CardRegisterPage> = {
         }}
       >
         <BrowserRouter>
-          <CardFormValueContext.Provider value={MOCK_CARD_CONTEXT}>
-            <Story />
-          </CardFormValueContext.Provider>
+          <CardFormProvider>
+            <CardFormValueContext.Provider value={MOCK_CARD_CONTEXT}>
+              <Story />
+            </CardFormValueContext.Provider>
+          </CardFormProvider>
         </BrowserRouter>
       </div>
     ),
@@ -39,10 +43,6 @@ const MOCK_CARD_CONTEXT: CardInfo = {
 };
 
 export default meta;
-type Story = StoryObj<typeof CardRegisterPage>;
+type Story = StoryObj<typeof CardNameRegisterPage>;
 
-export const Default: Story = {
-  args: {
-    registerCard: (cardData: CardData) => cardData,
-  },
-};
+export const Default: Story = {};
